+++ src/js/html5.js
@@ -82,6 +82,9 @@ const html5 = {
                 triggerEvent.call(player, player.media, 'qualitychange', false, {
                     quality: input,
                 });
+
+                // Save to storage
+                player.storage.set({ quality: input });
             },
         });
     },